User init provides options and an init function
Look for a specified filename, and load whatever namespace is in it. This allows multiple symbols
from the namespace to be used, rather than just loading it for side effects. For example, this init
file:

```
(ns init
  (:require [clijmage.viewer :as viewer]
            [clijmage.main :as main]
            [clijmage.images-coll :as images-coll]
            [clijmage.util :as util]))

(def startup-options
  {::main/load-image-coll-from-stdin false})

(defn init []
  (viewer/apply-bindings! {(main/combination javafx.scene.input.KeyCode/P [])
                           (util/runnable (fn [] (println (images-coll/current @main/images-position))))}))
```

Which shows that you can completely disable loading images based on stdin, if desired. You can even
disable running the follow-up init function. Although this seems strange, it allows init files to
skip defining it if they don't want to.