Separate per-image state out from the ordering and position state
Before this, the `images-position` ref was a forward-backward tracking a sequence of maps and the
`states-stack` ref was a stack of previous values from `images-position`. This model was extremely
elegant because all of the "current" state was in one data structure which could be replaced
wholesale (popping a value from the stack) or saved wholesale (pushing it into the stack and
creating a new one to make current).

However, this design could not express state which is shared "up and down". For example, there was
no way to narrow to the marked images and decide to unmark some! Refactoring the state is a bit less
elegant and loses some theoretical generality, but is more able to accurately express important use
cases. The new model keeps a map from path to the state map in the ref `image-states`, and keeps a
stack of forward-backwards representing the paths/sequences in the ref `sequence-stack`. This means
that making changes and such requires correlating data from the two structures, but it also means
that things can be pushed onto and popped from `sequence-stack` without impacting the states at all.