Replace narrow and sequence stack with show only marked
Instead of a stack of forward-backward objects, keep a single forward-backward object and a
predicate that can "hide" images. Narrowing to marked now means setting a predicate to hide any that
aren't marked.

This loses the expressive power of the stack (can't have multiple levels of narrowing), but avoids
the complexity of the same image path being in any number of forward-backward instances in the
stack. It is not clear if multiple levels of narrowing serve any purpose, but they can always be
addedd back by composing a new predicate that uses the old predicate.