Combine startup config and init fn
Move the user init fn earlier, and instead of reading a startup config var from the user init NS,
expect it to be returned by the init function. Pass in the commandline args.

This paves the way for the init fn to process arguments, if it chooses, and change the startup
options accordingly.