Make coll-state functions work from any thread
Previously, many could only be called from the Java FX Application Thread. This was because, despite
using thread-safe STM state storage, they called `viewer/goto!` directly, which manipulates the GUI
state.

This change adds `refresh-viewer!` which simply grabs the current state and calls viewer/goto!
either in the current thread (if the application thread) or dispatches a runnable to said thread
with `runLater`. While this is theoretically slightly less "precise" when multiple state updates
happen in rapid succession, the need for rapid GUI updates in that situation is highly tenuous.